refactor: rename feature
