+++ src/features/location-categories/help-requests/FilterHelpRequest.tsx
@@ -2,18 +2,18 @@ import { Filter } from "@/components/Filter/Filter";
 import { FilterHeader } from "@/components/Filter/FilterHeader";
 import { useTranslation } from "next-i18next";
 import { FilterControl } from "@/components/Filter/FilterControl";
+import { FilterOptions, TimeOption } from "@/utils/filterTime";
 import { MenuItem } from "@mui/material";
 import {
-  useServiceFilter,
-  filters as serviceFilters,
-  ServiceCategory,
-} from "../stores/useServiceFilter";
+  useHelpRequestFilter,
+  helpRequestFilters as helpRequestFilters,
+  HelpRequestCategory,
+} from "./useHelpRequestFilter";
 
-export const FilterService = () => {
+export const FilterHelpRequest = () => {
   const { t } = useTranslation("home");
-  const { isOpen, actions, selectedCategories } = useServiceFilter();
-
-  console.log({ selectedCategories });
+  const { selectedTime, isOpen, actions, selectedCategories } =
+    useHelpRequestFilter();
 
   if (!isOpen) {
     return null;
@@ -24,7 +24,7 @@ export const FilterService = () => {
       isOpen={isOpen}
       header={
         <FilterHeader
-          title={t("filter.servicesTitle")}
+          title={t("filter.helpRequestTitle")}
           onClose={() => {
             actions.setIsOpen(false);
           }}
@@ -36,18 +36,35 @@ export const FilterService = () => {
         value={selectedCategories}
         label={t("filter.reasonsTitle")}
         onChange={(_e, selectedCategories) => {
-          console.log({ selectedCategories });
           if (Array.isArray(selectedCategories)) {
             actions.setSelectedCategories(
-              selectedCategories as ServiceCategory[]
+              selectedCategories as HelpRequestCategory[]
             );
           }
         }}
       >
-        {Object.entries(serviceFilters).map(([key, option]) => {
+        {Object.entries(helpRequestFilters).map(([key, option]) => {
           return (
             <MenuItem key={key} value={option.type}>
               {t(`filter.category.${option.type}`)}
+            </MenuItem>
+          );
+        })}
+      </FilterControl>
+
+      <FilterControl
+        value={selectedTime}
+        label={t("filter.timestampTitle")}
+        onChange={(_e, [selected]) => {
+          if (typeof selected === "string") {
+            actions.setTimestamp(selected as TimeOption);
+          }
+        }}
+      >
+        {FilterOptions.map((option, idx) => {
+          return (
+            <MenuItem key={idx} value={option.value}>
+              {t(`filter.time.${option.label}`)}
             </MenuItem>
           );
         })}
